Removed the obsolete DataSetFilters module, added CropFilter as a separate module
