+++ src/main/java/net/sf/mzmine/modules/rawdatamethods/filtering/cropper/help/cropFilter.png
